services/users/requirements.txt
flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
flask-cors==3.0.3
flask-migrate==2.1.1
flask-testing==0.6.2

gunicorn==19.7.1

coverage==4.4.2
flake8==3.5.0
flask-debugtoolbar==0.10.1
